refactor(api): split isRoleValid and validRoles from user service
